Restrict Access to Data (Users of same company)

+++ BileMo-API/src/Controller/UserController.php
@@ -5,21 +5,24 @@ declare(strict_types=1);
 namespace App\Controller;
 
 use App\Entity\User;
+use App\Entity\Customer;
 use App\Repository\UserRepository;
 use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\Serializer\SerializerInterface;
+use Symfony\Component\Validator\Validator\ValidatorInterface;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Component\Serializer\Normalizer\AbstractNormalizer;
-use Symfony\Component\Validator\Validator\ValidatorInterface;
+use Symfony\Component\VarDumper\Cloner\AbstractCloner;
 
 /**
  * class UserController
  */
 #[Route('/api/users')]
-class UserController
+class UserController extends AbstractController
 {
     /**
      * Get all Users
@@ -33,8 +36,17 @@ class UserController
     #[Route(name: 'app_users_collection_get', methods:["GET"])]
     public function collection(UserRepository $userRepository, SerializerInterface $serializer): JsonResponse
     {
+        /**
+         * @var User $connectedUser
+         */
+        $connectedUser = $this->getUser();
+        
         return new JsonResponse(
-            $serializer->serialize($userRepository->findAll(), "json", ['groups' => 'get']),
+            $serializer->serialize(
+                $userRepository->findBy(["customer" => $connectedUser->getCustomer()]), 
+                "json", 
+                ['groups' => 'get']
+            ),
             JsonResponse::HTTP_OK,
             [],
             true
@@ -53,6 +65,18 @@ class UserController
     #[Route('/{id}', name: 'app_users_item_get', methods:["GET"])]
     public function item(User $user, SerializerInterface $serializer): JsonResponse
     {
+        /**
+         * @var User $connectedUser
+         */
+        $connectedUser = $this->getUser();
+        
+        if ($connectedUser->getCustomer() !== $user->getCustomer()){
+            return new JsonResponse(
+                null,
+                JsonResponse::HTTP_UNAUTHORIZED
+            );
+        }
+        
         return new JsonResponse(
             $serializer->serialize($user, "json", ['groups' => 'get']),
             JsonResponse::HTTP_OK,
@@ -81,23 +105,33 @@ class UserController
         UrlGeneratorInterface $urlGenerator,
         ValidatorInterface $validator
     ): JsonResponse {
+        /**
+         * @var User $connectedUser
+         */
+        $connectedUser = $this->getUser();
+        
         /** @var User $user */
         $user = $serializer->deserialize($request->getContent(), User::class, 'json');
         
         $errors = $validator->validate($user);
         if ($errors->count() > 0){
-            return new JsonResponse($serializer->serialize($errors,"json"), JsonResponse::HTTP_BAD_REQUEST);
+            return new JsonResponse(
+                $serializer->serialize($errors,"json"),
+                JsonResponse::HTTP_BAD_REQUEST
+            );
         }
 
+        $user->setCustomer($connectedUser->getCustomer());
         $em->persist($user);
         $em->flush();
 
         return new JsonResponse(
-            $serializer->serialize($user, "json", ['groups' => 'get']),
-            JsonResponse::HTTP_CREATED,
-            ["Location" => $urlGenerator->generate('app_users_item_get', ["id" => $user->getId()])],
-            true
-        );
+            $serializer->serialize(
+                $user, "json", ['groups' => 'get']),
+                JsonResponse::HTTP_CREATED,
+                ["Location" => $urlGenerator->generate('app_users_item_get', ["id" => $user->getId()])],
+                true
+            );
     }
 
     /**
@@ -119,6 +153,18 @@ class UserController
         EntityManagerInterface $em,
         ValidatorInterface $validator
     ): JsonResponse {
+
+        /**
+         * @var User $connectedUser
+         */
+        $connectedUser = $this->getUser();
+
+        if ($connectedUser->getCustomer() !== $user->getCustomer()){
+            return new JsonResponse(
+                null,
+                JsonResponse::HTTP_UNAUTHORIZED
+            );
+        }
 
         $user = $serializer->deserialize(
             $request->getContent(),
@@ -152,6 +198,17 @@ class UserController
     #[Route('/{id}', name: 'app_users_item_delete', methods:["DELETE"])]
     public function delete(User $user, EntityManagerInterface $em): JsonResponse
     {
+        /**
+         * @var User $connectedUser
+         */
+        $connectedUser = $this->getUser();
+
+        if ($connectedUser->getCustomer() !== $user->getCustomer()){
+            return new JsonResponse(
+                null,
+                JsonResponse::HTTP_UNAUTHORIZED
+            );
+        }
 
         $em->remove($user);
         $em->flush();
